backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pytest
httpx
google-genai
